requirements.txt
aiohttp==1.0.5
async-timeout==1.2.1
chardet==3.0.4
discord.py==0.16.8
flake8==3.3.0
greenlet==0.4.12
mccabe==0.6.1
msgpack-python==0.4.8
multidict==3.0.0
neovim==0.1.13
pycodestyle==2.3.1
pyflakes==1.5.0
websockets==3.3
